+++ src/app/locales/en/components/login-box/header.js
@@ -1,0 +1,5 @@
+export default {
+  brandTitle: 'Onezone',
+  brandSubtitle: 'development version',
+  isolatedZone: 'isolated zone',
+};