+++ src/app/locales/en/components/login-box/login-form-container.js
@@ -1,0 +1,15 @@
+export default {
+  formSubtitle: 'Using your Onepanel account',
+  formSubtitleTip: 'This login method is available for administrators and special users ' +
+    'created via onepanel administrative interface. Regular users must login via social or ' +
+    'institutional accounts, which can be found in previous menu.',
+  noProvidersFormSubtitleTip: 'This login method is available for administrators and special ' +
+    'users created via onepanel administrative interface. Regular users must login via social or ' +
+    'institutional accounts, which are currently disabled by the administrators of this zone.',
+  dropdownSubtitle: 'Pick your identity provider',
+  loginWith: 'Login with',
+  findProviderPlaceholder: 'Find your identity provider...',
+  fetchProvidersFailedUnknown: 'unknown reason',
+  authEndpointError: 'Getting authentication endpoint failed',
+  authEndpointConfError: 'Authorization configuration is invalid, please contact system administrator',
+};