+++ src/app/locales/en/components/login-box/social-box-list.js
@@ -1,0 +1,6 @@
+export default {
+  onepanelAccount: 'Onepanel account',
+  loginUsing: 'Login using ',
+  showMore: 'Show all identity providers',
+  fetchProvidersFailed: 'Server error - fetching login providers failed',
+};